+++ app/src/main/java/com/wngud/oneminutestart/OneMinuteApp.kt
@@ -7,7 +7,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
-import com.wngud.oneminutestart.presentation.BottomNavigationBar
+import com.wngud.oneminutestart.presentation.components.BottomNavigationBar
 
 @Composable
 fun OneMinuteApp() {
