+++ app/src/main/java/com/wngud/oneminutestart/presentation/statistics/StatisticsScreen.kt
@@ -9,7 +9,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
-import com.wngud.oneminutestart.presentation.AppBar
+import com.wngud.oneminutestart.presentation.components.AppBar
 
 @Composable
 fun StatisticsScreen(
